Simplify maven deps by using inheritance
Define dependency versions once in dependencyManagement and inherit in
child modules, to avoid defining the version and scope of modules more
than once.

Use version properties to make things even clearer.

Move the maven-bundle-plugin to the root of the project to avoid
repeating the same config.

Move the src/main/scala source configuration into the root by using the
build-helper-plugin.