Added Pax Exam based in-container integration test
New module integration-test that contains integration tests. Created an
initial test that obtains a reference to the DiningHakkerService, creates
a dining Hakker, and then interrogates its name and current state.

The dining Hakker messages were moved to the API bundle, so that the
integration test has access to them without seeing the implementation
classes in core.

The test is run via the standard "mvn install" command.